// apps/lab-app/src/types/onboarding/RegisterData.ts
export type RegisterData = {
    username: string;
    password: string;
    email: string;
    firstName: string;
    lastName: string;
    phone: string;
    address: string;
    city: string;
    state: string;
    zip: string;
    country: string;
    modules: number[];
    verified: boolean;
    // ===============
    profileImage: string;
    idProof: string;
    addressProof: string;
};